+++ src/lib.rs
@@ -11,6 +11,7 @@ pub mod constants;
 
 pub mod app;
 pub mod cli;
+pub mod clipboard;
 pub mod config;
 pub mod diagram;
 pub mod diff;